+++ packages/docs/src/stories/Switch.stories.tsx
@@ -1,0 +1,10 @@
+import type { StoryObj, Meta } from '@storybook/react'
+import { Switch, SwitchProps } from '@haruo-ui/react'
+
+export default {
+  title: 'Surfaces/Switch',
+  component: Switch,
+  args: {},
+} as Meta<SwitchProps>
+
+export const Primary: StoryObj<SwitchProps> = {}