+++ packages/react/src/components/Button/ButtonSelect.tsx
@@ -1,4 +1,4 @@
-import { styled } from '../styles'
+import { styled } from '../../styles'
 import { ComponentProps } from 'react'
 
 export const Button = styled('button', {