+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @haruo-ui/react
 
+## 3.1.0
+
+### Minor Changes
+
+- Add new component and fix button component
+- f6fb34f: Fix component slide size
+
 ## 3.0.0
 
 ### Major Changes